requirements.txt
pyleapcard==0.1.8
APScheduler==3.6.3
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
idna==2.9
joblib==0.16.0
lightgbm==2.3.1
numpy==1.19.1
pandas==1.0.5
pickle-mixin==1.0.2
PyMySQL==0.10.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scikit-learn==0.22.1
scipy==1.5.1
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
tzlocal==2.1
urllib3==1.25.9
